Safari-609.1.11.1/LayoutTests/editing/execCommand/deleting-text-rests-typing-style.html

<div id="description">This tests deleting a bold paragraph, then changing the selection and moving it back to the spot
where that paragraph was deleted and typing more. The new test should not be bolded.</div>
<div id="edit1" contentEditable="true"><div id="one"><b>Bold</b><br></div><div id="two"><br></div></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
Markup.description(document.getElementById('description').textContent);

one = document.getElementById("one");
window.getSelection().setBaseAndExtent(one, 0, one, one.childNodes.length);
document.execCommand("Delete");

window.getSelection().modify("move", "forward", "character");
window.getSelection().modify("move", "backward", "character");

document.execCommand("InsertText", false, "Not bold");

Markup.dump(document.getElementById('edit1'));
</script>
