LayoutTests/platform/ios-simulator/editing/execCommand/deleting-text-rests-typing-style-expected.txt

This tests deleting a bold paragraph, then changing the selection and moving it back to the spot
where that paragraph was deleted and typing more. The new test should not be bolded.
| <div>
|   id="one"
|   <br>
| <div>
|   id="two"
|   <br>
